chore: Move `~/modules/font` folder into `~/modules/customization`
