+++ mobile/src/db/schema.ts
@@ -14,7 +14,7 @@ import type { Prettify } from "~/utils/types";
 import type { PlayFromSource } from "~/stores/Playback/types";
 
 //? Need to export all schemas from a single file.
-export * from "~/modules/font/schema";
+export * from "~/modules/customization/font/schema";
 export * from "~/modules/theme/schema";
 
 //#region Artist
