+++ mobile/src/modules/customization/font/helpers/storage.ts
