+++ mobile/src/modules/customization/font/utils.ts
