+++ mobile/src/modules/scanning/hooks/useLoadResources.ts
@@ -11,7 +11,7 @@ import { checkForMigrations } from "../helpers/migrations";
 
 import { createImageDirectory } from "~/lib/file-system";
 import { Stopwatch } from "~/utils/debug";
-import { createFontDirectory } from "~/modules/font/helpers/storage";
+import { createFontDirectory } from "~/modules/customization/font/helpers/storage";
 
 /** Logic that should be run before rendering the onboarding/scanning progress views. */
 export function useLoadResources() {
