+++ mobile/src/navigation/routes.tsx
@@ -51,8 +51,8 @@ import ModifyTrack from "./screens/tracks/ModifyView";
 import Tracks from "./screens/tracks/View";
 import { TrackSheet } from "./screens/tracks/sheets/TrackSheet";
 import { ArtistsSheet } from "./sheets/ArtistsSheet";
+import FontScreenGroup from "~/modules/customization/font/screens";
 import EqualizerSettings from "~/modules/equalizer/screens/View";
-import FontScreenGroup from "~/modules/font/screens";
 import Search from "~/modules/search/screens/View";
 import ThemeScreenGroup from "~/modules/theme/screens";
